Now one can save not only set of prices but also a Average Overshoot length, plus overshoot only for up and for down

+++ src/MM.java
@@ -32,7 +32,7 @@ public class MM {
 
 
     public int computePriceShift(int exceedVolume){
-        double priceShift = alpha * (int)Math.sqrt(Math.abs(exceedVolume));
+        double priceShift = alpha * Math.sqrt(Math.abs(exceedVolume));
         if (exceedVolume < 0){
             return -(int) Math.round(priceShift);
         }
